added exception handler for 'EntityNotFoundException'. moved entities to they own package
