+++ src/main/java/pdm/persistence/model/entity/Category.java
@@ -1,4 +1,4 @@
-package pdm.persistence.model;
+package pdm.persistence.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
