+++ src/main/java/pdm/persistence/model/entity/FilmActorPK.java
@@ -1,4 +1,4 @@
-package pdm.persistence.model;
+package pdm.persistence.model.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
