+++ src/main/java/pdm/persistence/model/repository/FilmRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import pdm.persistence.model.Film;
+import pdm.persistence.model.entity.Film;
 
 import java.util.Set;
