+++ src/main/java/pdm/persistence/model/repository/LanguageRepository.java
@@ -2,7 +2,7 @@ package pdm.persistence.model.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import pdm.persistence.model.Language;
+import pdm.persistence.model.entity.Language;
 
 @Repository
 public interface LanguageRepository extends CrudRepository<Language, Long> {
