+++ src/main/java/pdm/persistence/web/service/CategoryToDTOService.java
@@ -2,8 +2,8 @@ package pdm.persistence.web.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
-import pdm.persistence.model.Category;
-import pdm.persistence.model.Film;
+import pdm.persistence.model.entity.Category;
+import pdm.persistence.model.entity.Film;
 import pdm.persistence.model.dto.CategoryDTO;
 import pdm.persistence.model.repository.CategoryRepository;
